Fixed duplicate SessionErrors class.
