+++ gymanager/src/main/java/com/faroc/gymanager/sessionmanagement/infrastructure/sessions/gateways/SessionsRepository.java
@@ -3,10 +3,10 @@ package com.faroc.gymanager.sessionmanagement.infrastructure.sessions.gateways;
 import com.faroc.gymanager.common.domain.exceptions.UnexpectedException;
 import com.faroc.gymanager.sessionmanagement.application.sessions.gateways.SessionsGateway;
 import com.faroc.gymanager.sessionmanagement.domain.common.time.TimeUtils;
+import com.faroc.gymanager.sessionmanagement.domain.sessions.errors.SessionErrors;
 import com.faroc.gymanager.sessionmanagement.domain.sessions.SessionReservation;
 import com.faroc.gymanager.sessionmanagement.domain.sessions.Session;
 import com.faroc.gymanager.sessionmanagement.domain.common.timeslots.TimeSlot;
-import com.faroc.gymanager.sessionmanagement.domain.sessions.errors.SessionErrors;
 import org.jooq.DSLContext;
 import org.jooq.codegen.maven.gymanager.tables.records.SessionsRecord;
 import org.springframework.stereotype.Repository;
