+++ gymanager/src/test/groovy/com/faroc/gymanager/sessionmanagement/unit/application/sessions/commands/AddSessionHandlerTests.groovy
@@ -8,12 +8,12 @@ import com.faroc.gymanager.sessionmanagement.application.sessions.gateways.Sessi
 import com.faroc.gymanager.sessionmanagement.application.trainers.gateways.TrainersGateway
 
 import com.faroc.gymanager.sessionmanagement.domain.rooms.Room
-import com.faroc.gymanager.sessionmanagement.domain.sessions.errors.SessionErrors
+
 import com.faroc.gymanager.common.domain.exceptions.ConflictException
 import com.faroc.gymanager.common.domain.exceptions.UnexpectedException
 import com.faroc.gymanager.sessionmanagement.domain.common.timeslots.TimeSlot
+import com.faroc.gymanager.sessionmanagement.domain.sessions.errors.SessionErrors
 import com.faroc.gymanager.sessionmanagement.domain.trainers.Trainer
-import com.faroc.gymanager.sessionmanagement.infrastructure.sessions.gateways.SessionCategoriesRepository
 import com.faroc.gymanager.sessionmanagement.unit.domain.rooms.utils.RoomsTestsFactory
 import com.faroc.gymanager.sessionmanagement.unit.domain.sessions.utils.SessionsTestsFactory
 import com.faroc.gymanager.sessionmanagement.unit.domain.trainers.utils.TrainersTestsFactory
@@ -77,8 +77,8 @@ class AddSessionHandlerTests extends Specification {
 
         then:
         def ex = thrown(UnexpectedException)
-        ex.getDetail() == SessionErrors.ROOM_NOT_FOUND
-        ex.getMessage() == SessionErrors.roomNotFound(roomId)
+        ex.getDetail() == SessionErrors.ADD_SESSION_ROOM_NOT_FOUND
+        ex.getMessage() == SessionErrors.addSessionRoomNotFound(roomId)
     }
 
     def "when adding session and trainer does not exist should throw unexpected exception"() {
